+++ server/migrations/20180813123653-create-order.js
@@ -15,16 +15,8 @@ module.exports = {
           key: 'id',
         },
       },
-      createdAt: {
-        allowNull: false,
-        defaultValue: Sequelize.fn('now'),
-        type: Sequelize.DATE,
-      },
-      updatedAt: {
-        allowNull: false,
-        defaultValue: Sequelize.fn('now'),
-        type: Sequelize.DATE,
-      },
+      createdAt: Sequelize.DATE,
+      updatedAt: Sequelize.DATE,
     })
   },
   down: (queryInterface, Sequelize) => {