+++ server/migrations/20180813123656-create-category.js
@@ -17,16 +17,8 @@ module.exports = {
       img: {
         type: Sequelize.TEXT,
       },
-      createdAt: {
-        allowNull: false,
-        defaultValue: Sequelize.fn('now'),
-        type: Sequelize.DATE,
-      },
-      updatedAt: {
-        allowNull: false,
-        defaultValue: Sequelize.fn('now'),
-        type: Sequelize.DATE,
-      },
+      createdAt: Sequelize.DATE,
+      updatedAt: Sequelize.DATE,
     })
   },
   down: (queryInterface, Sequelize) => {